modify: change files to use absolute imports
