+++ src/App.js
@@ -1,6 +1,5 @@
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
-import Home from './pages/Home';
-import User from './pages/User';
+import { Home, User } from 'pages';
 
 function App() {
   return (
