+++ src/components/Photo/index.js
@@ -1,12 +1,1 @@
-import { Link } from 'react-router-dom';
-import { Image } from './Photo.Styles';
-
-function Photo({ photo }) {
-  return (
-    <Link to={`/user/${photo.user.username}`}>
-      <Image src={photo.urls.thumb} alt={photo.description} />
-    </Link>
-  );
-}
-
-export default Photo;
+export { default } from './Photo';
