+++ src/components/PhotoList/index.js
@@ -1,16 +1,1 @@
-import Photo from '../Photo';
-import { List, ListItem } from './PhotoList.Styles';
-
-function PhotoList({ photos }) {
-  return (
-    <List>
-      {photos.map((photo) => (
-        <ListItem key={photo.id}>
-          <Photo photo={photo} />
-        </ListItem>
-      ))}
-    </List>
-  );
-}
-
-export default PhotoList;
+export { default } from './PhotoList';
