+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as Photo } from './Photo';
+export { default as PhotoList } from './PhotoList';
