+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export { default as Home } from './Home';
+export { default as User } from './User';
